pyproject.toml
[project]
name = "FetchGet"
version = "1.1.0"
description = "A free and open-source tool that help you to gather information of your system utilities and performing networking related tasks"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cython>=3.1.1",
    "discord-py>=2.5.2",
    "dnspython>=2.7.0",
    "httpx>=0.28.1",
    "ping3>=4.0.8",
    "psutil>=7.0.0",
    "pyfiglet>=1.0.3",
    "python-dotenv>=1.1.0",
    "python-whois>=0.9.5",
    "requests>=2.32.3",
    "rich>=14.0.0",
    "setuptools>=80.9.0",
    "speedtest-cli>=2.1.3",
    "statistics>=1.0.3.5",
    "vt-py>=0.21.0",
]
